requirements.txt
tensorflow
dvc
pandas
notebook
numpy
matplotlib
seaborn
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
types-PyYAML
types-requests
-e .